fix(hermes,router): decouple router-specific logic
