+++ libraries/hermes/src/Rest/Rest.ts
@@ -22,12 +22,23 @@ import { ConduitRoute } from '../classes';
 const swaggerUi = require('swagger-ui-express');
 
 export class RestController extends ConduitRouter {
+  private readonly _privateHeaders: string[];
   private _registeredLocalRoutes: Map<string, Handler | Handler[]>;
   private _swagger?: SwaggerGenerator;
 
-  constructor(grpcSdk: ConduitGrpcSdk, swaggerRouterMetadata: SwaggerRouterMetadata) {
+  constructor(
+    grpcSdk: ConduitGrpcSdk,
+    privateHeaders: string[] = [],
+    swaggerRouterMetadata: SwaggerRouterMetadata = {
+      urlPrefix: '',
+      securitySchemes: {},
+      globalSecurityHeaders: [],
+      setExtraRouteHeaders(): void {},
+    },
+  ) {
     super(grpcSdk);
     this._registeredLocalRoutes = new Map();
+    this._privateHeaders = privateHeaders;
     this._swagger = new SwaggerGenerator(swaggerRouterMetadata);
     this.initializeRouter();
   }
@@ -158,8 +169,7 @@ export class RestController extends ConduitRouter {
           if (r.redirect) {
             res.removeHeader('Authorization');
             res.removeHeader('authorization');
-            res.removeHeader('clientid');
-            res.removeHeader('clientsecret');
+            this._privateHeaders.forEach(h => res.removeHeader(h));
             return res.redirect(r.redirect);
           } else {
             let result;
