+++ libraries/hermes/src/index.ts
@@ -2,7 +2,7 @@ import express, { NextFunction, Request, Response, Router } from 'express';
 import { RestController } from './Rest';
 import { GraphQLController } from './GraphQl/GraphQL';
 import { SocketController } from './Socket/Socket';
-import ConduitGrpcSdk, { ConduitError, Indexable } from '@conduitplatform/grpc-sdk';
+import ConduitGrpcSdk, { ConduitError } from '@conduitplatform/grpc-sdk';
 import http from 'http';
 import {
   ConduitRequest,
@@ -16,43 +16,6 @@ import cookieParser from 'cookie-parser';
 import path from 'path';
 import { ConduitRoute } from './classes';
 import { createRouteMiddleware } from './utils/logger';
-
-const swaggerRouterMetadata: SwaggerRouterMetadata = {
-  urlPrefix: '',
-  securitySchemes: {
-    clientId: {
-      name: 'clientid',
-      type: 'apiKey',
-      in: 'header',
-      description: 'A security client id, retrievable through [POST] /security/client',
-    },
-    clientSecret: {
-      name: 'clientSecret',
-      type: 'apiKey',
-      in: 'header',
-      description:
-        'A security client secret, retrievable through [POST] /security/client',
-    },
-    userToken: {
-      type: 'http',
-      scheme: 'bearer',
-      bearerFormat: 'Bearer',
-      description:
-        'A user authentication token, retrievable through [POST] /authentication/local or [POST] /authentication/renew',
-    },
-  },
-  globalSecurityHeaders: [
-    {
-      clientId: [],
-      clientSecret: [],
-    },
-  ],
-  setExtraRouteHeaders(route: ConduitRoute, swaggerRouteDoc: Indexable): void {
-    if (route.input.middlewares?.includes('authMiddleware')) {
-      swaggerRouteDoc.security[0].userToken = [];
-    }
-  },
-};
 
 export class ConduitRoutingController {
   private _restRouter?: RestController;
@@ -70,6 +33,7 @@ export class ConduitRoutingController {
     private readonly baseUrl: string,
     private readonly grpcSdk: ConduitGrpcSdk,
     cleanupTimeoutMs: number = 0,
+    private readonly privateHeaders: string[] = [],
     private readonly swaggerMetadata?: SwaggerRouterMetadata,
   ) {
     this._cleanupTimeoutMs = cleanupTimeoutMs < 0 ? 0 : Math.round(cleanupTimeoutMs);
@@ -121,7 +85,8 @@ export class ConduitRoutingController {
     if (this._restRouter) return;
     this._restRouter = new RestController(
       this.grpcSdk,
-      this.swaggerMetadata ?? swaggerRouterMetadata,
+      this.privateHeaders,
+      this.swaggerMetadata,
     );
   }
